+++ addons/yat/src/types/NetworkingOptions.cs
@@ -6,4 +6,5 @@ public record NetworkingOptions
 	public ushort TTL { get; init; } = 30;
 	public ushort BufferSize { get; init; } = 32;
 	public bool DontFragment { get; init; } = true;
+	public ushort Delay { get; init; } = 1000;
 }